requirements_production.txt
# AutoFund AI - Production Dependencies

# Core LLM
anthropic>=0.34.0

# FastAPI & Production Backend
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.8

# Redis & Task Queue
redis>=5.0.1
rq>=1.15.1
celery>=5.3.4

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-decouple>=3.8

# API Rate Limiting & Security
slowapi>=0.1.9
helmet>=0.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
newrelic>=9.2.0

# HTTP Client
httpx>=0.25.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Production
gunicorn>=21.2.0
flower>=2.0.1